Exits seeding process gracefully and reorganizes files
